+++ src/main/java/com/example/deliverytracker/order/dto/OrderResponse.java
@@ -2,9 +2,9 @@ package com.example.deliverytracker.order.dto;
 
 import com.example.deliverytracker.order.entity.Order;
 import com.example.deliverytracker.store.dto.StoreResponse;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -14,13 +14,13 @@ public class OrderResponse {
 
     private Order.Status status;
 
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     private StoreResponse store;
 
     private List<OrderItemResponse> orderItems;
 
-    private OrderResponse(Long id, Order.Status status, int totalPrice, StoreResponse store, List<OrderItemResponse> orderItems) {
+    private OrderResponse(Long id, Order.Status status, BigDecimal totalPrice, StoreResponse store, List<OrderItemResponse> orderItems) {
         this.id = id;
         this.status = status;
         this.totalPrice = totalPrice;
